实现了精英敌机，超级精英敌机，Boss敌机，多种道具，有音效，有menu界面

+++ src/edu/hitsz/basic/AbstractFlyingObject.java
@@ -17,23 +17,23 @@ public abstract class AbstractFlyingObject {
     /**
      * x 轴坐标
      */
-    protected int locationX;
+    protected double locationX;
 
     /**
      * y 轴坐标
      */
-    protected int locationY;
+    protected double locationY;
 
 
     /**
      * x 轴移动速度
      */
-    protected int speedX;
+    protected double speedX;
 
     /**
      * y 轴移动速度
      */
-    protected int speedY;
+    protected double speedY;
 
     /**
      * 图片,
@@ -63,7 +63,7 @@ public abstract class AbstractFlyingObject {
     public AbstractFlyingObject() {
     }
 
-    public AbstractFlyingObject(int locationX, int locationY, int speedX, int speedY) {
+    public AbstractFlyingObject(double locationX, double locationY, double speedX, double speedY) {
         this.locationX = locationX;
         this.locationY = locationY;
         this.speedX = speedX;
@@ -104,8 +104,8 @@ public abstract class AbstractFlyingObject {
         int fFactor = flyingObject instanceof AbstractAircraft ? 2 : 1;//对方
 
         //对方坐标、宽度、高度
-        int x = flyingObject.getLocationX();
-        int y = flyingObject.getLocationY();
+        int x = (int)flyingObject.getLocationX();
+        int y = (int)flyingObject.getLocationY();
         int fWidth = flyingObject.getWidth();
         int fHeight = flyingObject.getHeight();
 
@@ -115,11 +115,11 @@ public abstract class AbstractFlyingObject {
                 && y - ( fHeight/fFactor+this.getHeight()/factor )/2 < locationY;
     }
 
-    public int getLocationX() {
+    public double getLocationX() {
         return locationX;
     }
 
-    public int getLocationY() {
+    public double getLocationY() {
         return locationY;
     }
 
@@ -128,7 +128,7 @@ public abstract class AbstractFlyingObject {
         this.locationY = (int) locationY;
     }
 
-    public int getSpeedY() {
+    public double getSpeedY() {
         return speedY;
     }
